# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Inv-MQTT Multi-Level Channel Platform API
  description: |
    光伏逆变器监控系统 - 多级渠道平台 API 参考文档
    
    ## 核心功能
    
    - **组织管理**: 制造商、代理商、分销商、客户、服务商的多级组织架构
    - **邀请系统**: 用户邀请与自动注册
    - **设备认领**: 设备唯一性验证与所有权转移
    - **成员生命周期**: 成员添加、移除、转移的完整流程
    
    ## 认证方式
    
    所有端点 (除特别标注) 需要使用 JWT Bearer Token 认证。Token 通过登录接口获取。
    
    ## 错误响应
    
    所有错误响应遵循统一格式:
    ```json
    {
      "code": "ERROR_CODE",
      "message": "错误描述",
      "details": {} // 可选的详细信息
    }
    ```
  
  version: 1.0.0
  contact:
    name: CSERGY Smart Energy Platform Support
    email: support@csergy.com
  license:
    name: Proprietary
    url: https://csergy.com/terms
servers:
  - url: https://api.inv-mqtt.com/api/v1
    description: 生产环境
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: https://dev-api.inv-mqtt.com/api/v1
    description: 测试环境
...
paths:
...
  /organizations:
    post:
      summary: 创建组织单元
      tags: [Organizations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
            example:
              name: "华南分公司"
              type: "distributor"
              parent_id: null
      responses:
        '201':
          description: 组织创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationWithChildren'
        '400':
          description: 请求参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 禁止创建 (最终用户权限不足)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 冲突 (重复名称或层级限制)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        
    get:
      summary: 列出组织单元
      tags: [Organizations]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码 (从 1 开始)
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          description: 每页数量 (最大 100)
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: type
          in: query
          description: 按类型筛选
          required: false
          schema:
            type: string
            enum:
              - manufacturer
              - agent
              - distributor
              - customer
              - service_partner
        - name: status
          in: query
          description: 按状态筛选
          required: false
          schema:
            type: string
            enum:
              - active
              - disabled
              - quarantined
      responses:
        '200':
          description: 组织列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrganizationList'
        '403':
          description: 禁止访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token 在登录后获取，有效期 15 分钟
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 错误码
          example: "FORBIDDEN"
        message:
          type: string
          description: 人类可读的错误描述
          example: "Forbidden access"
        details:
          type: object
          description: 额外的错误详情 (可选)
...
    CreateOrganizationRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "华南分公司"
        type:
          type: string
          enum:
            - manufacturer
            - agent
            - distributor
            - customer
            - service_partner
          example: "distributor"
        parent_id:
          type: integer
          format: int64
          nullable: true
          description: 父组织 ID，null 表示直接隶属于根租户
...
    OrganizationType:
      type: string
      enum:
        - manufacturer
        - agent
        - distributor
        - customer
        - service_partner
    OrganizationStatus:
      type: string
      enum:
        - active
        - disabled
        - quarantined
    OrganizationSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          $ref: '#/components/schemas/OrganizationType'
        status:
          $ref: '#/components/schemas/OrganizationStatus'
        has_children:
          type: boolean
          description: 是否有子节点
    OrganizationWithChildren:
      type: object
      properties:
        id:
          type: integer
          format: int64
        root_tenant_id:
          type: integer
          format: int64
        parent_id:
          type: integer
          format: int64
          nullable: true
        type:
          $ref: '#/components/schemas/OrganizationType'
        code:
          type: string
          description: 组织编码 (可选)
        name:
          type: string
        status:
          $ref: '#/components/schemas/OrganizationStatus'
        version:
          type: integer
          format: int64
          description: 乐观锁版本号
        children_count:
          type: integer
          description: 子节点数量
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        child_organizations:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationSummary'
    PaginatedOrganizationList:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationWithChildren'
        total:
          type: integer
          format: int64
        page:
          type: integer
        page_size:
          type: integer
...
security:
  - bearerAuth: []
...